# servants-of-india/docs/openapi.swagger.yaml
...
openapi: 3.0.3
info:
  title: Servants of Bharat API — Complete (All APIs)
  description: Swagger / OpenAPI 3.0 documentation for ALL APIs created for the Servants of Bharat platform.
    Every operation includes its mapped user story (x-user-story) and error-handling details (x-error-handling).
  version: 1.0.0
servers:
- url: /api
  description: Local API base path
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: 'Forbidden: insufficient role'
...
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
          example: tree-plantation-a1b2c3d4
        title:
          type: string
        description:
          type: string
        category_id:
          type: string
          format: uuid
        category_name:
          type: string
        venue:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        event_date:
          type: string
          format: date
        start_time:
          type: string
          example: 09:00
        end_time:
          type: string
          example: '13:00'
        capacity:
          type: integer
          nullable: true
        status:
          type: string
          enum:
          - upcoming
          - completed
          - cancelled
        created_by:
          type: string
          format: uuid
        created_by_name:
          type: string
    EventInput:
      type: object
      required:
      - title
      - description
      - category_id
      - venue
      - address
      - city
      - state
      - event_date
      - start_time
      - end_time
      properties:
        title:
          type: string
        description:
          type: string
        category_id:
          type: string
          format: uuid
        venue:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        event_date:
          type: string
          format: date
        start_time:
          type: string
          example: 09:00
        end_time:
          type: string
          example: '13:00'
        capacity:
          type: integer
          nullable: true
        status:
          type: string
          enum:
          - upcoming
          - completed
          - cancelled
...
    Certificate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        volunteer_id:
          type: string
          format: uuid
        volunteer_name:
          type: string
        certificate_number:
          type: string
          example: SOB-2026-00001
        verification_code:
          type: string
        pdf_url:
          type: string
        issued_at:
          type: string
          format: date-time
...
paths:
...
  /events:
    get:
      tags:
      - Events
      summary: List events
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
      security:
      - bearerAuth: []
      parameters:
      - name: category
        in: query
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: string
      x-user-story: As any user, I want to browse events so that I can decide where to serve.
      x-error-handling:
      - 401 - not authenticated
      description: 'List events


        **User story:** As any user, I want to browse events so that I can decide where to serve.


        **Error handling:**

        - 401 - not authenticated'
    post:
      tags:
      - Events
      summary: Create event (Event Manager / Super Admin)
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Event'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': *id001
        '403': *id002
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInput'
      x-user-story: As an event manager, I want to create events so that I can run service activities.
      x-error-handling:
      - 422 - missing required fields or invalid category_id
      - 403 - caller is not an Event Manager / Super Admin
      - 401 - not authenticated
      description: 'Create event (Event Manager / Super Admin)


        **User story:** As an event manager, I want to create events so that I can run service activities.


        **Error handling:**

        - 422 - missing required fields or invalid category_id

        - 403 - caller is not an Event Manager / Super Admin

        - 401 - not authenticated'
  /events/{identifier}:
    get:
      tags:
      - Events
      summary: Get event by UUID or slug
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Event'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: string
      x-user-story: As any user, I want to view an event's details so that I can learn more about it.
      x-error-handling:
      - 404 - event not found
      - 401 - not authenticated
      description: 'Get event by UUID or slug


        **User story:** As any user, I want to view an event''s details so that I can learn more about
        it.


        **Error handling:**

        - 404 - event not found

        - 401 - not authenticated'
    put:
      tags:
      - Events
      summary: Update event (creator or Super Admin)
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Event'
        '403': *id002
        '401': *id001
      security:
      - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInput'
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: string
      x-user-story: As an event manager, I want to edit my own events so that I can keep them accurate.
      x-error-handling:
      - 403 - not the event's creator (and not Super Admin)
      - 404 - event not found
      - 422 - invalid field/category
      description: 'Update event (creator or Super Admin)


        **User story:** As an event manager, I want to edit my own events so that I can keep them accurate.


        **Error handling:**

        - 403 - not the event''s creator (and not Super Admin)

        - 404 - event not found

        - 422 - invalid field/category'
    delete:
      tags:
      - Events
      summary: Delete event (creator or Super Admin)
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '403': *id002
        '401': *id001
      security:
      - bearerAuth: []
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: string
      x-user-story: As an event manager, I want to delete my own events so that I can remove cancelled
        ones.
      x-error-handling:
      - 403 - not the event's creator (and not Super Admin)
      - 404 - event not found
      description: 'Delete event (creator or Super Admin)


        **User story:** As an event manager, I want to delete my own events so that I can remove cancelled
        ones.


        **Error handling:**

        - 403 - not the event''s creator (and not Super Admin)

        - 404 - event not found'
...
  /certificates/generate:
    post:
      tags:
      - Certificates
      summary: Generate my certificate (idempotent)
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Certificate'
        '400':
          description: Not all categories completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': *id001
        '403': *id002
      security:
      - bearerAuth: []
      x-user-story: As a volunteer, I want to generate my certificate once all five categories are complete
        so that I have proof of service.
      x-error-handling:
      - 400 - all five categories must be completed first
      - 401 - not authenticated
      - 403 - caller is not a Volunteer
      description: 'Generate my certificate (idempotent)


        **User story:** As a volunteer, I want to generate my certificate once all five categories are
        complete so that I have proof of service.


        **Error handling:**

        - 400 - all five categories must be completed first

        - 401 - not authenticated

        - 403 - caller is not a Volunteer'
  /certificates/me:
    get:
      tags:
      - Certificates
      summary: Get my certificate
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Certificate'
        '404':
          description: No certificate yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': *id001
        '403': *id002
      security:
      - bearerAuth: []
      x-user-story: As a volunteer, I want to view my certificate so that I can download it.
      x-error-handling:
      - 404 - no certificate issued yet
      - 401 - not authenticated
      - 403 - caller is not a Volunteer
      description: 'Get my certificate


        **User story:** As a volunteer, I want to view my certificate so that I can download it.


        **Error handling:**

        - 404 - no certificate issued yet

        - 401 - not authenticated

        - 403 - caller is not a Volunteer'
  /certificates:
    get:
      tags:
      - Certificates
      summary: List all certificates (Super Admin)
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '401': *id001
        '403': *id002
      security:
      - bearerAuth: []
      x-user-story: As a super admin, I want the record of all issued certificates so that I can audit
        recognition.
      x-error-handling:
      - 401 - not authenticated
      - 403 - caller is not a Super Admin
      description: 'List all certificates (Super Admin)


        **User story:** As a super admin, I want the record of all issued certificates so that I can audit
        recognition.


        **Error handling:**

        - 401 - not authenticated

        - 403 - caller is not a Super Admin'
  /certificates/{id}/download:
    get:
      tags:
      - Certificates
      summary: Get certificate download URL
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '401': *id001
        '403': *id002
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      x-user-story: As a certificate holder, I want a download link so that I can save my certificate.
      x-error-handling:
      - 404 - certificate not found
      - 403 - not the owner (volunteers)
      - 401 - not authenticated
      description: 'Get certificate download URL


        **User story:** As a certificate holder, I want a download link so that I can save my certificate.


        **Error handling:**

        - 404 - certificate not found

        - 403 - not the owner (volunteers)

        - 401 - not authenticated'
  /verify/{verification_code}:
    get:
      tags:
      - Certificates
      summary: 'PUBLIC: verify a certificate'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
      parameters:
      - name: verification_code
        in: path
        required: true
        schema:
          type: string
      x-user-story: As anyone (no login), I want to verify a certificate by its code so that I can trust
        it is genuine.
      x-error-handling:
      - 200 - always returns {valid:true/false}; public, no auth required
      description: 'PUBLIC: verify a certificate


        **User story:** As anyone (no login), I want to verify a certificate by its code so that I can
        trust it is genuine.


        **Error handling:**

        - 200 - always returns {valid:true/false}; public, no auth required'
...